refactor goldmaster test code to SearchAndReplaceArbitrary
I've refactored this bit into it's own module and switched to using
node's fs/promises API to read/write to the json file. I did this to
ensure that the file we were using for the goldmaster data was consistent
on both reading and writing and to make the usage of the arbitrary code
in that module a bit more apparent.